+++ src/Features/CSharp/Portable/RemoveUnusedVariable/CSharpRemoveUnusedVariableCodeFixProvider.cs
@@ -3,10 +3,10 @@
 using System.Collections.Immutable;
 using System.Composition;
 using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CodeFixes.RemoveUnusedVariable;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.RemoveUnusedVariable;
 
-namespace Microsoft.CodeAnalysis.CSharp.CodeFixes.RemoveUnusedVariable
+namespace Microsoft.CodeAnalysis.CSharp.RemoveUnusedVariable
 {
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = PredefinedCodeFixProviderNames.RemoveUnusedVariable), Shared]
     [ExtensionOrder(After = PredefinedCodeFixProviderNames.AddImport)]
@@ -15,6 +15,7 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeFixes.RemoveUnusedVariable
         private const string CS0168 = nameof(CS0168);
         private const string CS0219 = nameof(CS0219);
 
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(CS0168, CS0219);
+        public sealed override ImmutableArray<string> FixableDiagnosticIds 
+            => ImmutableArray.Create(CS0168, CS0219);
     }
 }
